requirements.txt
# LangChain Core
langchain
langchain-core

# groq
langchain-groq

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Graph for chains
grandalf
